feat: implementación de expedientes médicos con blindaje RLS y reparación de historial de migraciones

+++ supabase/migrations/20260425180000_matricula_ciclo_schema.sql
@@ -73,7 +73,7 @@ insert into public.alumno_ciclo (alumno_id, ciclo_id, grado, grupo, grupo_id)
 select
   a.id,
   c.id,
-  a.grado,
+  a.grado::integer,
   a.grupo,
   g.id
 from public.alumnos a
